flask_app/requirements.txt
connexion >= 2.6.0
connexion[swagger-ui]==2.14.2
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
click==8.0.4
Flask==2.0.3
gunicorn==20.1.0
itsdangerous
Jinja2==3.0.3
MarkupSafe
Werkzeug==2.0.3
datadog
opentelemetry-distro
opentelemetry-instrumentation-wsgi
opentelemetry-sdk
opentelemetry-api
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp